Multi-Modal-Deep-Learning-Based-Digital-Restoration-of-Ancient-Paintings/requirements.txt
tensorflow
keras
pillow
numpy
opencv-python
flask
